add open_port to dimmer object, remove special dashboard's object views

+++ db/priv/repo/migrations/20181214172735_add_thermometers.exs
@@ -11,7 +11,7 @@ defmodule DB.Repo.Migrations.AddThermometers do
 
     create table(:therm_temp_reads) do
       add :therm_id, references(:thermometers, on_delete: :delete_all)
-      add :value, :integer
+      add :value, :float
       timestamps()
     end
